backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.2
pymongo==4.6.0
